feat: Updates logic to use the repositoryUrl from semantic release
BREAKING CHANGE:
This most likely won't be a true breaking change to anyone as it now aligns with the way the semantic-release/git plugin pushes the initial tag (using the repositoryUrl provided by semantic-release).

Previously, it was just pushing to 'origin' which worked fine when there was a git.config, which usually is the case when using the actions/checkout action, but could cause errors when it was not set (e.g. when using the option ` persist-credentials: false`).